Move date and time string splitting to _date-format.js
We can use `getDateAndTimeStringsFromDate()` to split a `Date` object
into a date string and a time string.
